refactor: extract VerdictRenderer and finish MainActivity render decomposition
